[git] add a third option that allows to get combined diff for merge commits
In f8a956f5140db0f0fcda2f4aba25228dea525389 an option was introduced to get no diff for merge commits. In a457eeef9bff95658896d5019732f4667cd468f7 is was made default by mistake, while actually combined diff if expected. For example, GitPushOperationSingleRepoTest was broken because of this change.  In oder to fix the problem a third option is added for combined diff and used by default.