More robustly disable the Maven plugin
We have Studio-specific code in PluginManagerCore which disables
the Maven plugin. This CL changes that code so that, instead of 'disabling'
the Maven plugin, we prevent PluginManagerCore from seeing that it even exists.

This more closely resembles the semantics in a release build.
It happens to help fix some test failures in ASwB in which the tests
manipulate the 'disabled plugins' list directly.

Bug: n/a
Test: existing
Change-Id: Iba317a492b187203df44c65cff415ff59aaf8975