Force load Kotlin even if wrong platform prefix
Our bundled Kotlin plugin has a hard dependency on
com.intellij.modules.androidstudio, meaning that it only
loads if the platform prefix is set to 'AndroidStudio'.

However, some tests and run configurations run with platform prefix
'Idea', and we would still like to load the Kotlin plugin in those
cases. This CL forces that to happen by (1) ignoring Kotlin's dependency
on com.intellij.modules.androidstudio, and (2) ignoring the Maven plugin
even when running as IDEA.

Context for (2): the Kotlin plugin has an optional dependency on the
Maven plugin, and will load extra stuff if Maven is there. But we strip
the bundled Kotlin plugin of that extra stuff. So things break if the
Maven plugin and the Kotlin plugin are enabled at the same time.

Test: existing
Bug: 130328344
Change-Id: Icc4730496b9e5d91e05b4634b2d27be6ec43e9a1