Decrement the current index after removing first element from buffer
There is an issue in Android Studio where the build output parsers are
not receiving all errors. The reason to that is when we try to read the
next line while the buffer is full, the first element is removed and the
current index will be pointing to the next element so this line will be
skipped in the next call of readLine.

This change decrements the current index after the first element is
removed from the buffer so it will point to the same element before
the removal.