Unstage file for deletion if it's being reused
Currently in SchemeManagerImpl it is possible for a scheme's containing
file to be deleted AFTER another scheme of the same containing file name
starts using the file.

SchemeManagerImpl stages files to be deleted but doesn't actually do so
until save() is called. While a file is staged for deletion, another
scheme can be "soft" saved with saveScheme(), which saves the scheme in
memory. saveScheme() does contain code to unstage the scheme's
containing file for deletion, but there is an edge case where this won't
happen due to an early exit that assumes no operation is required to
"soft" save the scheme if the new scheme is identical to the old one.
This results in a zero net-change in schemes, and a file staged for
unintentional deletion.

This moves the "unstage file for deletion" code before the exit so
it's always checked, which fixes the edge case scenario.

(cherry picked from commit 6ed6a6a)

Bug: 127677541
Change-Id: I0e6c2df09b2b5654325dc8ae9aa3a07c8914fd1d