bundle netty-handler-proxy
netty-codec-socks-4.1.34.Final.jar 119 KB
netty-handler-proxy-4.1.34.Final.jar 22 KB

This year TeamCity plugin will be migrated to REST API and opposite to Upsource will use http client that based on Netty (highly likely). But problem is that since 2018.2 IntelliJ Platform doesn't bundle netty proxy handler classes anymore. Now it is fixed (probably will be cherry-picked to 191).

It is tricky question should 3rd-party plugin bundle own plugin version or not, but in case of http client it is not only about reducing plugin download size and final class path, but about reusing resources (in case of netty - threads/buffers/so on).