[vcs-log-graph] make sure that edges color matches the color of one of the nodes
Color of the node is either determined by corresponding head or by its layout index. Before this commit, color of the edge matched the color of its up node, but not all the time: when up and down node layout indexes were different, greater layout index was used. This lead to the edge being different from both nodes, which really stood out and made the graph look more complex and colorful than it was. In this commit, color of one of the nodes is selected all the time, and the choice of the node is determined by layout index.