vcs: load contents for the Merge dialog under a common modal progress
MergeProvider#loadRevisions used to be called synchronously from the EDT, but all implementations were wrapping the loading process into VcsUtil.runVcsProcessWithProgress. It is more logical for loadRevisions() to be synchronous, and to have a single common modal progress above all of them.

Luckily, all 3-party plugins call the same runVcsProcessWithProgress() method, so this commit patches it to avoid wrapping into a modal progress (and thus failing with an exception), if the method is called from a pooled thread.

The method itself has little sense now, so it marked as deprecated. A few usages not from loadRevisions() are updated to use good old ProgressManager.run*().

Relates to IDEA-208359.

GitOrigin-RevId: 519f0f691c0b01a1405ab563086c2ab7daab64e9