[file-history] track renames through a merge commit
When file history stops at a merge commit, we need to check changes to all of the parent commits to find a possible move. In order to do so, "-m" parameter should be passed to the `git show` command. With this parameter the command returns several records instead of one. The number of returned records is the same as the number of merged parents, since this commit would not be present in file history with the added name-status otherwise.

GitOrigin-RevId: a41fc7934d041205687ed1b957bd99ead83dc6ed