[BOW] Select the most useful error message
If at least one error is parsed by one of the parsers, the first one will be selected
and the "run" node will be collapsed.
otherwise all the error nodes will be expanded and the root will be selected because it
has a good chance to contain the most useful message.

Bug: 130722131
Test: Manually
Change-Id: Ie0a33997a77a0fe358dfef0cfe6c3b7e82e33951