vcs: merge dialog: load the multiple files description label asynchronously: IDEA-208359
The GitDefaultMergeDialogCustomizer asks Git about the branch which is being merged, etc. However, preloading this information is difficult to implement due to numerous and different usages of the customizer. So we just load this information right from the dialog, and display a stub text until the necessary information is ready (usually it is very fast).

GitOrigin-RevId: 6e6cf45d16ec2071c583c133f5305744c5e52367