Option added not to flush text in test runners until new line.
By default, OutputEventSplitter flushes any arrived text except when in the middle of service message.

So, when test outputs "a" client gets "a".
Some clients need text to be buffered until new line char.

Such clients want "a" to stay buffered until test runner adds "\n" to it.

OutputEventSplitter now supports it, and OutputToGeneralTestEventsConverter configures it to do so, unless ConsoleProperties marked as editable, see javadoc in OutputToGeneralTestEventsConverter

GitOrigin-RevId: 8c000a1717067d77bffb7c224f58ae70e282e9f0