[directory-history] add a hack that allows to see directory history before rename
This hack introduces a special provider that can read renames from vcs.xml, and an implementation of FileHistoryData that uses renames from this provider. Directories that are considered renamed need to be marked as added or deleted in affected commits map first. This way finding renames is triggered for them.

IDEA-211416

GitOrigin-RevId: 04605bcbb5a69ec452a8f7827b82d3e9e3b988cc