Fix IDE support for git via ssh user/password prompt
IJ migrated from SSHMain to GitNativeAskPassApp (d6b37824238b9efc3e9e0e0876386869c60aa409)
as the git helper to get username & passwords from the UI into the git command line.

This code was not yet using the Android Studio specific Authentication logic to the
builtin-webserver.

As our builtin-webserver modifications require token based authentication for all requests,
this broke the IPC commands used by GitNativeAskPassApps to prompt the user for username/password.

The fix is to pass the GIT_ASKPASS_TOKEN variable for all GitNative commands and pick that
variable up from the git helper and inject it as a header when talking to the IDE builtin-webserver.

Fixes: 124680708
Test: manual

Change-Id: If2756533e7e035c18dcc1405e7ca6364469cf8d4