[vcs-log] do not return changelists from MainFrame to prevent commit details from loading
When VcsRevisionNumber is asked from context, and not found in the target component, a VcsRevisionNumberArrayRule is invoked. This rule tries to get the revision from changelists, triggering loading of selected commits. VcsRevisionNumber is actually available, but in one of the parent components (VcsLogPanel), but the rule is called before going there.

 Changelists were introduced in 02d7048bf28a3822dd70ba9d4c008760cd7c88d0 to support revision history from the Log, however are no longer needed for this purpose. So they can be removed.

IDEA-214490, IDEA-212138

GitOrigin-RevId: 180dfea919f720eb49d08d9e4de323c4a8a81335