git: cleanup: remove synchronized marker from GitHandler methods
The class is not thread-safe: it is intended to be populated and started within a single thread, and existing synchronized modifiers don't help with thread safety anyway.

At the same time, output and error output are populated from other threads (readers of the external process), but they are stored in synchronized collections, while synchronizing the start() method is not helping here anyway.

GitOrigin-RevId: af9fbef893dee88924e8798e5c02369563ee282b