[directory-history] add a really ugly hack to make refine algorithm work with filters
In order to correctly track file names in the graph, current implementation of the refine algorithm needs to have commits with renames present in the visible graph. This is not the case if such commits do not match active filters. In order to fix this, commits with renames are added to the graph temporarily and removed from it after refine is finished.

IDEA-211416

GitOrigin-RevId: 5f550c611919f295e5e669f5cb2bd5fffff2dbd0