build scripts: explicitly specify main module before module with patched dependencies file in Python plugins layout
This is needed to fix building searchable options index without ProductModulesLayout.mainModules. In that case order of modules in classpath will be determine by order they are specified in PluginLayout, and we need to ensure that the main module comes before the buildPatchesModule. Otherwise version of python-plugin-dependencies.xml from the buildPatchesModule will be used, and plugins included in it will be loaded twice, so building searchable options index will fail.

GitOrigin-RevId: 0ae1e833983bf9dc732e57d4d1254b88a090fa27