java: extract java support to a separate plugin in IntelliJ IDEA (IDEA-195719)
All java-related components are removed from the platform part of IntelliJ IDEA. This also introduces 'intellij.java.ide.resources' module for resources which should be included into IntelliJ IDEA and other Java-based IDEs but shouldn't be included into the Java plugin itself. Currently the same plugin.xml is used for the both community and ultimate editions, but some additional JARs are included into the plugin distribution for Ultimate Edition. The plugin is marked as 'essential-plugin' in IntelliJ IDEA and Android Studio so it won't be visible in Settings and cannot be disabled.

Change-Id: I6fe5c64133672aae09b743bda8a007558f1f6791