builds scripts: use restricted compatible build range when publishing plugins for EAP builds
Before this change when a plugin is published with EAP or release build its 'until-build' was set to xxx.*. It means that the plugin will be compatible with the next product EAP or release build, so user will be able to update the IDE without updating the plugin. This could lead to compatibility problems if API is changed in newer IDE build and also plugin's users won't report feedback about the new plugin version. Now when publishing plugins with a EAP build we use restricted compatible build range (e.g. 'until-build' is set to 'xxx.yyy.*') so the plugin will not be compatible with the next EAP or release and users will need to update it together with the IDE.

GitOrigin-RevId: bc52b18b4bcf324b5b601f2a1c2585bbe155621c