GO-5838 Local history not working for non-".go" files when more than one project is open
The problem occurred when there was several open projects with intersecting content roots:
in one project a folder was excluded, while in another it was a source folder.

It was because we considered a file not versioned if any open project had it excluded, that was not correct.
To fix, we now consider a file unversioned, if it's excluded from all open project.

Note, that 'ignored files' are per-application, so they are always not versioned.

GitOrigin-RevId: a071cf0b5652fbfa68159f7aa68099b982606b7c