PY-30493 Update, Uninstall, etc. actions for Python packages
PyCharm delegates Python packaging actions to the Python packaging tools:
pip, pipenv, conda. It simply show the progress of these external commands
and then updates the file system if there are any changes in the project
files or the project interpreter files.

The list of actions that should be available:

  * Install (already available)
  * Upgrade (an IntelliJ platform action)
  * Uninstall (an IntelliJ platform action)
  * Use Conda Package Manager (for Conda environments)
  * Show Early Releases

GitOrigin-RevId: a1887a19fcab4e104bdfc7dd2af36082c5320c44