Support options without assigned value in git config
Usually options in git config have values, but there are options that can be used without a value, e.g. required flag of a filter. By default, ini4j expects every option to have a value, and thus reading of a git config might fail. The fix is to enable support of empty values in ini4j. This fixes IDEA-215904

GitOrigin-RevId: 3a38b62ea169a4434143eecb2789cac417321094