[file-history] return correct file path for visible pack with paths information in additional data
When history tab is set invalid, a snapshot visible pack is created to temporary store information. When user switches back to the tab, the snapshot pack is shown until the tab is refreshed. The snapshot has paths information, but it's not an instance of FileHistoryVisiblePack. So a correct FilePath needs to be returned from the ordinary VisiblePack.

EA-145252

GitOrigin-RevId: 3c4fb6711513a38ffa03fb9849b452635aaa71b0