git: don't save and don't load branch filters in the Update Info tab
These could be accidentally set by a user, e.g. to "HEAD", since it is easy to confuse the Update Info tab with an ordinary Log tab. And then the next Update Project will show "100500 commits matching filter", where 100500 is the number of all commits reachable from the HEAD. IDEA-216661

GitOrigin-RevId: c6f6d57337ceda1b397254db5adda5286939cae5