vcs-log: support hash-range in multi-repository projects: IDEA-216785
To filter by a range, we need to resolve the supplied ref to a CommitId. To speedup resolve for a full-size hash, no checks were performed: the string was simply converted to a Hash and a CommitId. Therefore in a multi-root project a full-size hash was resolved in all roots, although it existed only in one of them.

To fix this, check if the generated CommitId exists in the storage. If it doesn't, then we generated it incorrectly, and there is no such hash in the root.

GitOrigin-RevId: 3b34a6212e539355a0e0f40ed5ade9160ad855d9