vcs-log: support hash-range in multi-repository projects when filtering via VCS, support ranges by any ref expression
Since the previous commit the hashes are resolved correctly when filtering by graph or the index. However, if we need to filter in VCS (e.g. because there are other filters, beside the range filter), all ranges were passed to all providers and all roots. Therefore a log provider could request commits filtered by a range of commits not existing in this repository.

This continues the fix for IDEA-216785.

To solve this, teach the VcsLogProvider to resolve provided reference and convert it to a commit hash, if the reference is known to this provider in this root. If the reference is not resolved in a root, don't ask for commits in this root.

As a nice side effect, any refs known to Git can now be entered as range limits, for example, tags: they will become unresolved by the Log engine, but will be resolved by the GitLogProvider.

GitOrigin-RevId: f09347a0992ce0f63e5117200d84fdf3cfafdcbe