git: remember the previous HEAD position before pull/merge
The revision is used to calculate updated files. It used to work, because Repository.getCurrentRevision returns the cached version which is not updated unless explicitly requested and unless a VFS refresh is received indicating that the HEAD file has changed: the first didn't happen, the latter most likely didn't happen as well, but it is just by luck.

GitOrigin-RevId: a44a65db4167dc15a459b481702f62cb0c187519