git: don't wait for the log refresh to show the Update Project notification if custom filters are not set
For simplicity, consider only the Path filter: it will be moved to the Settings in one of next commits, and will be the only preserved filter.

If there is no Paths filter, we don't need the up-to-date VisiblePack to show the notification. Therefore collect only the necessary information (available at once from Git), and show the notification. If user presses the "View Commits" link fast, he'll have to wait until the log is refreshed and filtered.

If there is a Path filter, then we'll have to wait, because the number of commits matching the filter is unavailable until the VisiblePack is ready.

Partial fix of IDEA-216137.

GitOrigin-RevId: d55d22d2faecf5c844fcced7c2639d3f1ed101aa