IDEA-215432: UnstableApiUsage\ScheduledForRemoval inspections: check if containing declaration is marked with annotation
If a class is marked with @ApiStatus.Internal annotation, all its methods must be considered internal, even if invocation is performed via a subclass and no explicit reference onto a class exists.
If a package is marked with @ApiStatus.Internal, all its classes and their members must be considered internal.

GitOrigin-RevId: f8bcb2a91cf958b878cd45a58d4fed69297fd372