[file-history] do not show "Commit not found" notification on opening file history
When opening file history for file or folder, either the first commit or some ancestor of the target commit (for "History Up to Here") should be selected. But when this is impossible to do, error notification should not be displayed as user did not explicitly requested to see this commit.

(cherry picked from commit b157f79195278adce6244e04b0cd869396c03b9c)

GitOrigin-RevId: cf51d6a25462eaac9e171265dd1d656ec7344de7