[vcs-log] dispose Log before Project
In this commit dispose is moved to ProjectManagerListener#projectClosing(), which is called before disposing the project. In order to do dispose the Log correctly, we first unsubscribe from mappings changes, then start dispose task and close the executor. After that, waiting for executor to terminate is done under modal progress (which can not be used from dispose() since it does not start a background thread under write action). In case someone manages to call createLog() after disposeLog() but before executor shutdown, a special flag myDisposeStarted is used to check whether disposing is in progress.

IDEA-217010, IDEA-219575

GitOrigin-RevId: 90ea64cb1fc99d07302320cd5ab11ae645170805