Merge ScheduledForRemovalInspection into UnstableApiUsageInspection.
ScheduledForRemovalInspection is effectively the same as more general UnstableApiUsageInspection. The only thing that differs is the provided message "Foo is scheduled for removal in 2.3.4" instead of "Foo is marked unstable". To customize the message a dedicated "message provider" interface is extracted. It may later be utilized for other annotations.

Also tests for "scheduled for removal API usages" are moved next to the tests of the UnstableApiUsageInspection.

GitOrigin-RevId: 0693f41954bd5a4d79699b83a76cd4d293a56f9a