Don't ignore exceptions that occur after models are fetched.
This change fixes an issue where exceptions that occured after the
models have been fetched (during task running) were not handed back to
the GradleProjectResolverExtensions.

In this change we pass all GradleConnectionExceptions that occurs during
the Gradle build to the resolvers via the buildFinished(Exception)
callback. This will be a super set of the exceptions that can be thrown
during fetching the modesl (as exceptions from tasks are not included
there).

Note: For Gradle versions <1.8 we don't pass the exception to the
buildFinished callback, since there it is not possible to run tasks via
the tooling API in these versions this is safe.

Signed-off-by: Nikita Skvortsov <nikita.skvortsov@jetbrains.com>

GitOrigin-RevId: ae7c6db0a985e438093ac860bd32b8c3c546e3e8