Optimize async stack frame loading
For each async stack frame we have to find a com.sun.jdi.Location
corresponding to a given line number. To do this we called
ReferenceType.locationsOfLine()---but that calls locationsOfLine() on
all methods in the class. Since we already know the name of the method
we're looking for, it's faster to first filter by method name, then
call locationsOfLine() directly on the matching methods.

This optimization is less important when running on a local JVM,
but very useful for remote debugging sessions. For example,
when debugging on an Android device, this change reduces the time spent
in findLocation() by 10x when loading 15 async stack frames, and the
frames show up in the UI significantly faster.

Before: 2767 ms spent in findLocation()
After: 247 ms spent in findLocation()