vcs: install listeners on the EDT: IDEA-217202
If the CommandListener is installed not on a EDT, then there are chances that it is installed in the middle of some command execution.
In this case we don't receive the commandStarted() event, but we do receive the commandFinished() event.
Because of this myCommandLevel is initially set to -1 and never reaches zero, thus we only collect myAdded/Deleted/etc/Files, and never perform the actual addition/removal/etc procedure.

To fix this, install listeners on EDT. Since Commands are executed on the EDT without a break, that ensures that we get both commandStarted and commandFinished.

GitOrigin-RevId: d66512ad30fcfe1c6f75b82c199c62ec575eea74