Revert "vcs: install listeners on the EDT: IDEA-217202"
This reverts commit d66512ad

Although, the command is executed on the EDT, and commandStarted/commandFinished are sent on the EDT as well, there is no guaranteed (and is actually not true) that all these operation will happen within a single EDT frame. So installing the listener on the EDT doesn't fix anything. A correct fix will follow.

GitOrigin-RevId: 6cc9761cf4bb175c822d47ff41281ebd4262fac4