vcs: don't count the depth level of the currently executed command
The commandFinished() event is sent only for top-level commands, so to execute a code after a stack of commands is completed, it is enough to do it in commandFinished().

This fixes the race IDEA-217202: if the CommandListener is installed between commandStarted() and commandFinished(), the code inside commandFinished() would never be executed because the counter was always negative.

This fix replaces the previous incorrect attempt made in d66512ad

(cherry picked from commit d8212cf31c89cc9e9dfe13f4dead89ef0504164e)

GitOrigin-RevId: d275d2006329327ff16851103ec2e8b75823407f