java: add explicit bridge methods to access protected members inherited by an enclosing class from a platform class
If a platform class and a plugin class derived from it are located in the same package, and an inner or local class in the plugin class access protected members declared in the plugin class, javac won't generate bridge methods so the code will fail with IllegalAccessError if the plugin class and the platform class are loaded by different class-loaders. Such cases are now properly reported by "Suspicious package-private access" inspections (see IDEA-217306) and here required bridge methods are manually introduced.

This is needed to switch off 'use-idea-classloader' option in Java plugin.

GitOrigin-RevId: b54cbafaaf407b493622b9deece5ea1ca2beedee