RUBY-24557 don't assign a potentially null scope to `FindUsagesOptions#searchScope` (annotated with `@NotNull`)
39afd1b31d80ece84dd0675eb4abfbc9f9ee63df introduced a race condition for populating and pre-selecting the search scope
in the combo box. For IDEs that cache the find usages options, this can cause the search scope to be `null` and saved,
resulting in an NPE later on down the track. This fix just ensures we can't set `null` to a `@NotNull` field, but
probably it would still be good to fix the actual issue where pre-selection should wait until population has completed.

GitOrigin-RevId: f5a1be3e9f0b93bb4b64c2cf4640fd01a985668b