Open links in the editor on mouseReleased event
Due to the fact that unfolding folded regions is happening on mouseReleased event,
opening links on mouseClicked leads to unfortunate behavior with unfolding:
if the link is under the folded region, a single click on the region unfolds it and opens the link simultaneously.

Moving opening link to mouseReleased event allows avoiding this situation.
Additional handling initial mousePressed event is supposed to filter out drag events on links.

GitOrigin-RevId: ce69d1b6041c4c70795c930b2ae15d0e518fdb97