IDEA-CR-49748: java: extract Java-specific part of Java-based IDEs to a separate plugin
There are some classes related to Java support which shouldn't be included to Java plugin when it's installed into another IDE (e.g. IntelliJIdeaLanguageCustomization) and these classes depend on classes from Java plugin. Since platform cannot depends on a plugin we need to extract such classes and related resources to a separate com.intellij.java.ide plugin (which is marked as essential-plugin and therefore won't be visible in Settings and cannot be disabled). This is needed to get rid of deprecated option use-idea-classloader in Java plugin which will make it a true plugin (i.e. it'll be able to depend on other plugins, see IDEA-195719).

This is a preparation commit, necessary files are moved without changes to save Git history in them.

(cherry picked from commit 2295a4ab872a63c9e11f88ac99d3418d06b96aa8)

GitOrigin-RevId: 7bdf5016ea1ec624086f44ee82fc0d8ac41d1a5d