Expose TextEditorWithPreview actions to subclasses
Subclasses of TextEditorWithPreview might want to customize the
presentation of ToggleActions used to change the editor mode. That can
be achieved by exposing the actions creation to subclasses through
protected methods. Subclasses, in turn, can create their own actions,
using TextEditorWithPreview's as delegates.

An advantage of this approach is also providing subclasses the ability
to create listeners for actions selection, selecting an action
programatically, etc.