git: use "Rollback" instead of "Revert" for the action which discards the local uncommitted changes
Revert means another thing and Git and using it for this action can be confusing.

There is no good name for this operation in native Git, because the operation consists of two parts - reset and checkout - and both names have multiple meanings and should be avoided.

Rollback seems the best choice: it doesn't have any other meaning in Git, and thus can't be confused with some other operation, it is used in some other VCSs.

IDEA-149015 UX-783

GitOrigin-RevId: 096224c91a730f45d349dcc212ce82d1cebd06dd