git: allow to change the path filter in the Update Info tab: IDEA-218028
If the Paths filter is not set (as it is initially), take the value from the Settings. Specifically handle the case when it is reset to All (when the filter is null).

Don't persist filters: store them only for the current Update Info tab.

(cherry picked from commit 40528c73aa62c3396606c368c7069b629c8bfae6)

GitOrigin-RevId: 8bdad68ead7a450a0a385cdd94dbbe67f165087f