PY-36737 Fix inability to use Jupyter editor actions from toolbar shown in popup
The toolbar shown in popup does not have the editor as its ancestor in Swing component hierarchy. This leads to the problem that neither `Editor` nor corresponding `VirtualFile` could be retrieved from `DataContext` passed to the toolbar actions. To solve this problem `PopupToolbar` implements `DataProvider` and uses `getDataContext()` from the outer `ActionToolbarImpl` which has these values.

GitOrigin-RevId: 8c000df10898a795fd700e1458dc8de90dca2377