Add perfgate_multi_run tag to perfgate bazel target
This bazel target produces data which is supposed to be uploaded to
perfgate, but currently the data isn't being uploaded because the
bazel target isn't being run by the perfgate_linux AB target.
Adding the perfgate_multi_run tag will cause the bazel target
to be run by the perfgate_linux AB target.

Bug: 136658973
Test: existing
Change-Id: I83e4f33d56e18a1bb7a31fa8f610286ea6a9a1b0