PY-7711 PY-36912 Don't spell check strings containing injections
rather delegating to the spell checker of the injected language.

Otherwise, we either just report typos twice (PY-36912) or, worse, report
numerous false positives since we try to apply tokenization rules for Python
to a different language as we notoriously have been doing all this time for
injected regular expressions.

Implementations of spell checkers for other languages such as Java or
JavaScript have been using this policy for ages.

GitOrigin-RevId: d1a2a3d48b7a5efff0dbb32ec2847fba1482f3a8