[vcs-log] call invokeAndWait() from the executor service to init or dispose ui
In order to init log, VcsLogManager is created in background thread, and then ui is initialized in EDT. Disposing works in the opposite way: first ui is disposed in EDT, then VcsLogManager is disposed in background. To execute init/dispose operations sequentially, all operations are submitted to the executor that uses a single thread, and working with ui is done from invokeAndWait().

Since it's not recommended to use invokeAndWait inside of the synchronized block, VcsLogManager is moved to the volatile field. This makes it impossible to ensure that method addProjectLogListener() works correctly, on the other hand this provides an alternative to the fix of the deadlock with PredictionService (IDEA-216389, f895741e18201990757aed75cb08a165b023f1b1, e9acfbf059baed6bdc69e6b8340c630178030ddd). This commit removes addProjectLogListener() method, and PredictionService just checks whether the listener was added twice.

GitOrigin-RevId: 48c7f22c67de293a3ac34dcb7bde06249b8f14ff