[vcs-log] display log filtering errors in empty text
Instead of keeping previous visible pack when filtering error occurs, create an instance of special class ErrorVisiblePack with an empty graph and an exception instance. Handle error visible packs similarly to error data packs: display error text with an action link. But in case of filtering errors, suggest user to either reset the filters or to filter again.

IDEA-193806

GitOrigin-RevId: 7353811f3fa1ce39ec5fe439ecf7c8b848799e9c