[stats-collector] Log the same factors values as we feed to the ML model
There was a possibility when the ML model sees the different factors comparing to factors being logged. This is possible because there is a delay between the ranking performing and the factors collecting. Now we make a snapshot of the used factors and then log only this snapshot instead of recalculated values. Making of a factors snapshot allows us to collect 'inside-session' features. This change also should improve the quality of the logs.

Also changes:
* Introduce a storage in a lookup for collecting of ml-completion related data (like features, completion settings, performance impact, etc).
* Localize session-factors related data in two instances: ElementSessionFactorsStorage and LookupSessionFactorsStorage.
* Store user factors in the new storage instead of directly in the lookup userdata.

GitOrigin-RevId: a48d5e84d7d5f7f66a860164d771d7c97fc42b20