UpdateCheckerService: do not run in unit test mode
Previously, the check would happen on appFrameCreated() in
AppLifecycleListener, which meant never for unit tests (headless).

After the change UpdateCheckerComponent -> UpdateCheckerService
in 2019.2, the update check runs unconditionally as a post-startup
activity (and in a hermetic build, we also lose the fact that it ran
once because we don't persist the config/options/updates.xml).

Bug: N/A
Test: manual
Change-Id: I9822bdbae06e7ff30fd5d5a3f5fae8ee9fcb325d