[vcs-log] count indexing tasks correctly
In order to determine when index for a vcs root is ready and could be used for filtering, a count of all running or scheduled indexing requests is kept for each root in myNumberOfTasks field. Before this commit, this count was incorrectly calculated: incremented when in fact a request was not created, and not decremented when a request was created (so its value was bigger). This happened either when vcs root was added to big repositories list (affects users who resumed paused indexing) or when scheduled commits list was empty (does not usually happen), and resulted in index not being used for filtering.

(cherry picked from commit ad9a0f2f1287b7a2bd5910e294ba944f059ba1d7)

GitOrigin-RevId: 7cf02ff2b975b71e3c0db9095dab0abbe541f69c