git: setup authentication prompts for all commands, not only for remote ones
When LFS is used, checking out a branch, or performing any other "local" command which requires fetching content from the repository, may require downloading this content from the LFS server => we must be ready for an authentication prompt even for local commands. This shouldn't affect performance or have any side effects, so it seems to be safe just to set up authentication prompts for all commands.

IDEA-217990, IDEA-218129

GitOrigin-RevId: af4757b29346e2046f8622a18d989d6d9f6c64c6