Define git config gc.auto for the repository according to team-wide shared settings
The default gc.auto setting is not efficient for some big projects (including IntelliJ IDEA itself), because such projects produce a lot of garbage, and git gc gets called too often. It makes sense to redefine the setting for such repositories. Since Git doesn't have a mechanism for sharing git config values, let's use intellij.yaml for this task.

If there is no explicit gc.auto value in the local .git/config, then set it to the value defined for the project in intellij.yaml.

GitOrigin-RevId: 2756489489263d23c0d8e7e2975044f68345b68c