schedule builtin server to start up later
This basically allows for netty to be loaded in a non-blocking way.

Implementation note.

The following contract is still preserved: if a command line
is passed to IDEA instance, it is either processed by a new instance, or is sent to
an already running instance.

Previously, SocketLock has had a clear `underLocks` structured
call which was meant to guarantee atomicity of creating a new builtin server;
it was needed to ensure singleton property of IDEA. If two instances are run, one
of them will acquire the locks first and start up builtin server; the other will
wait for locks to be up and then be able to phone builtin server of the first instance.

Unfortunately, we can't just postpone builtin server creation because we have to be able
to process activation cmdline from subsequent instances. For the sake of that, the locks
are being held "much longer": they are being get in a non-structured way and given up
only when netty is initialized, which happens later in an another thread.

GitOrigin-RevId: c36af0ec99a5775c2c4ef089d166796ae94577e5