IntelliJ project configuration: use Ant 1.7.1 as a dependency when compiling intellij.java.rt module
The shared Ant project library was updated to a version which requires Java 8, but intellij.java.rt has language level 1.3 and compiled with Java 1.6. This works when javac from Java 8 is used for compilation, but when a module with JDK 11 is added to IntelliJ project build process starts to use javac from JDK 11 as the main one and compile intellij.java.rt module using a forced javac from 1.6 and compilation fails.

GitOrigin-RevId: 2f30734e5594b13e14421d3d462e7d0755df086d