[vcs-log] move current branch condition calculation to background thread
Calculating current branch conditions in EDT made Log repaint slow when highlighting current branches was on. This commit moves calculation to the background thread using an ExecutorService. Each condition is wrapped into a ConditionHolder, which allows to keep conditions state with it. When a new data pack arrives, condition is not dropped as before, but marked "outdated". When an "outdated" condition is requested next time from the client code, a calculation is started and a condition is marked as "in progress". This allows to request calculation only when needed.

The new code is thread safe, so "isDispatchThread" assertions are removed.

IDEA-217125

GitOrigin-RevId: b7767877e6e0adbcd4e508640b6c736e2d381d74