PY-26264: Always put the caret to the end of the console text
Sometimes the caret was placed at the beginning of the last line (to combat IDEA-124688) -
however, this looks broken when further user input is expected.

Let's always move the caret all the way to the end, and then adjust scrolling position,
and then adjust the scroll position for the "scroll to the end" feature (as it's already
done in IDEA-184208).

GitOrigin-RevId: b609ca8d8a844e5c48263da30a84f99f6f1240dc