PY-34019 Test generation of skeletons for multiple modules and progress indication via JSON
I added a new command line option "--name-pattern" for the new generator3 mode
where it attempts to generate skeletons for all discovered modules that allows
to specify a glob pattern for module names. It was needed to restrict its output
in tests but it might also become handy for GenerateBinaryStubFix where we have
to perform similar filtering on Java side at the moment.

GitOrigin-RevId: 1dae4782a84f160f0d35ae8dda3b24ae01bfee82