[change-reminder] do not create content revisions while processing commits
This commit adds a method to GitCommit that allows accessing paths affected by this commit directly. When changes in the commit aren't parsed yet, this allows to avoid expensive and unnecessary creation of GitContentRevision instances.

FileHistoryProvider only takes one FilePath from each commit, but since there are no renames, this is equivalent to taking both paths, which is what the new method does.

IDEA-219351

GitOrigin-RevId: 71425add1af27c47d8ae140bd3bd9f06944db8d1