[git] create new factory in "View Commits" action
The factory keeps log manager, range filter and a listener that allows to calculate new commits. Log manager can be disposed, so the reference to it should not be saved. Listener is necessary for showing a correct notification and is useless after that. Range filter is already passed as parameter. So the instance of MyLogUiFactory is not needed.

GitOrigin-RevId: 98295e92e12b699e53a3799333af992075ef0f98